requirements.txt
Flask==2.2.2
pandas==1.5.3
mysql-connector-python==8.0.33
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.7
numpy>=1.21.0
